Fixed type typo on chebyshev filter bounds

+++ include/quda.h
@@ -666,10 +666,10 @@ extern "C" {
     int filter_iterations_between_vectors[QUDA_MAX_MG_LEVEL];
 
     /** Conservative estimate of largest eigenvalue of operator used for Chebyshev filter setup */
-    int filter_lambda_min[QUDA_MAX_MG_LEVEL];
+    double filter_lambda_min[QUDA_MAX_MG_LEVEL];
 
     /** Lower bound of eigenvalues that are not enhanced by the initial Chebyshev filter */
-    int filter_lambda_max[QUDA_MAX_MG_LEVEL];
+    double filter_lambda_max[QUDA_MAX_MG_LEVEL];
 
     /** Null-space type to use in the setup phase */
     QudaNullVectorSetupType setup_type[QUDA_MAX_MG_LEVEL];
